Planning: move path_bounds_decider into a folder.
