+++ modules/planning/tasks/BUILD
@@ -38,7 +38,7 @@ cc_library(
         "//modules/planning/proto:planning_config_proto",
         "//modules/planning/tasks/deciders:decider",
         "//modules/planning/tasks/deciders:path_assessment_decider",
-        "//modules/planning/tasks/deciders:path_bounds_decider",
+        "//modules/planning/tasks/deciders/path_bounds_decider:path_bounds_decider",
         "//modules/planning/tasks/deciders/path_lane_borrow_decider",
         "//modules/planning/tasks/optimizers/dp_st_speed:dp_st_speed_optimizer",
         "//modules/planning/tasks/optimizers/open_space_trajectory_generation:open_space_trajectory_provider",
