+++ modules/planning/tasks/deciders/BUILD
@@ -57,29 +57,11 @@ cc_library(
     copts = ["-DMODULE_NAME=\\\"planning\\\""],
     deps = [
         ":decider_base",
-        ":path_bounds_decider",
         ":path_decider_obstacle_utils",
         "//modules/planning/common:planning_context",
         "//modules/planning/common:planning_gflags",
         "//modules/planning/common:reference_line_info",
-    ],
-)
-
-cc_library(
-    name = "path_bounds_decider",
-    srcs = [
-        "path_bounds_decider.cc",
-    ],
-    hdrs = [
-        "path_bounds_decider.h",
-    ],
-    copts = ["-DMODULE_NAME=\\\"planning\\\""],
-    deps = [
-        ":decider_base",
-        ":path_decider_obstacle_utils",
-        "//modules/planning/common:planning_context",
-        "//modules/planning/common:planning_gflags",
-        "//modules/planning/common:reference_line_info",
+        "//modules/planning/tasks/deciders/path_bounds_decider:path_bounds_decider"
     ],
 )
 
@@ -165,18 +147,6 @@ cc_test(
     ],
     deps = [
         ":open_space_roi_decider",
-        "@gtest//:main",
-    ],
-)
-
-cc_test(
-    name = "path_bounds_decider_test",
-    size = "small",
-    srcs = [
-        "path_bounds_decider_test.cc",
-    ],
-    deps = [
-        "path_bounds_decider",
         "@gtest//:main",
     ],
 )
